requirements.txt
python-dotenv==1.0.0
langchain==0.0.154
streamlit==1.28.0
faiss-cpu==1.7.4
streamlit-extras
altair<5
openai==0.28.1
tiktoken==0.5.1
